+++ db/models/advertisement.go
@@ -8,7 +8,7 @@ type Advertisement struct {
 	User      User   `gorm:"foreignKey:UserID;constraint:OnUpdate:CASCADE,OnDelete:CASCADE"`
 	Text      string `gorm:"type:text"`
 	ImageID   string `gorm:"size:255;default:''"`
-	Status    uint
+	Status    uint8
 	CreatedAt time.Time
 	UpdatedAt time.Time
 }
